chore: address review comments
