+++ crates/miden-standards/asm/standards/note/mod.masm
@@ -1,3 +1,42 @@
+# miden::standards::note
+#
+# Collects note-related helper procedures that operate on input notes and are candidates for
+# eventual promotion into the protocol layer (`miden::protocol`).
+
+use miden::protocol::active_note
+use miden::protocol::input_note
+
 pub mod execution_hint
 pub mod note_creator
 pub mod note_id
+
+# PROCEDURES
+# =================================================================================================
+
+#! TODO: Move this to miden::protocol::input_note.
+#!
+#! Loads the storage items of the input note at `note_index` into memory at `dest_ptr`.
+#!
+#! The items are read from the advice map keyed by the note's storage commitment and validated
+#! against it, so a host that supplies wrong storage aborts the transaction.
+#!
+#! Inputs:  [note_index, dest_ptr]
+#! Outputs: []
+#!
+#! Where:
+#! - note_index is the index of the input note whose storage is loaded.
+#! - dest_ptr is the memory address the storage items are written to.
+#!
+#! Panics if:
+#! - the note index is greater or equal to the total number of input notes.
+#! - the advice-provided storage does not match the note's storage commitment.
+#!
+#! Invocation: exec
+pub proc input_note_get_storage
+    exec.input_note::get_storage_info
+    # => [NOTE_STORAGE_COMMITMENT, num_storage_items, dest_ptr]
+
+    # note that this call does NOT depend on the active note, it is merely a helper
+    exec.active_note::write_storage_to_memory
+    # => []
+end
