+++ crates/miden-testing/tests/scripts/mod.rs
@@ -5,6 +5,7 @@ mod code_inspection;
 mod expiration;
 mod faucet;
 mod faucet_policy_action;
+mod fee_collection;
 mod fee_manager;
 mod fee_sponsorship;
 mod non_fungible_faucet;
